requirements.txt
gcloud==0.17.0
googleapis-common-protos==1.6.0
httplib2==0.14.0
jws==0.1.3
oauth2client==3.0.0
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycrypto==2.6.1
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.13.0
